time: move primitives from core to structure/time
- Files were oddly placed in core. This fixes that.
