+++ src/main.cpp
@@ -1,12 +1,9 @@
 #include "spine/controller/pid.hpp"
 #include "spine/controller/sr_latch.hpp"
-#include "spine/core/datetime.hpp"
 #include "spine/core/debugging.hpp"
 #include "spine/core/exception.hpp"
 #include "spine/core/meta/enum.hpp"
 #include "spine/core/meta/unique_type_variant.hpp"
-#include "spine/core/schedule.hpp"
-#include "spine/core/timers.hpp"
 #include "spine/core/utils/concatenate.hpp"
 #include "spine/core/utils/string.hpp"
 #include "spine/core/utils/time_repr.hpp"
@@ -35,6 +32,9 @@
 #include "spine/structure/ringbuffer.hpp"
 #include "spine/structure/stack.hpp"
 #include "spine/structure/static_string.hpp"
+#include "spine/structure/time/datetime.hpp"
+#include "spine/structure/time/schedule.hpp"
+#include "spine/structure/time/timers.hpp"
 #include "spine/structure/tuple.hpp"
 #include "spine/structure/units/si.hpp"
 #include "spine/structure/vector.hpp"
