+++ src/spine/controller/pid.hpp
@@ -4,7 +4,7 @@
 #include "spine/controller/implementations/pid/pid_tuner.hpp"
 #include "spine/core/debugging.hpp"
 #include "spine/core/logging.hpp"
-#include "spine/core/timers.hpp"
+#include "spine/structure/time/timers.hpp"
 
 #include <cmath>
 #include <functional>
